Day 25 - US States Game
Day 24 of #100daysofcode

Today we went back to the Snake Game on day 20/21, and we created a text file for storing our high scores.

Then we finalized with a Letter merge program where we were given a list of names to send invitation letters to. Instead of creating a new personified letter per person, we used a generic letter and replaced the names with the names of our invitees.